+++ hypothesis-python/tests/numpy/test_argument_validation.py
@@ -276,22 +276,6 @@ def test_raise_invalid_argument(function, kwargs):
         function(**kwargs).example()
 
 
-@nps.defines_dtype_strategy
-def bad_dtype_strategy():
-    return st.just([("f1", "int32"), ("f1", "int32")])
-
-
-@given(st.data())
-def test_bad_dtype_strategy(capsys, data):
-    s = bad_dtype_strategy()
-    with pytest.raises(ValueError):
-        data.draw(s)
-    val = s.wrapped_strategy.mapped_strategy.value
-    assert capsys.readouterr().out.startswith(
-        "Got invalid dtype value=%r from strategy=just(%r), function=" % (val, val)
-    )
-
-
 @checks_deprecated_behaviour
 @given(st.data())
 def test_byte_string_dtype_len_0(data):
@@ -304,6 +288,12 @@ def test_byte_string_dtype_len_0(data):
 def test_unicode_string_dtype_len_0(data):
     s = nps.unicode_string_dtypes(min_len=0, max_len=0)
     assert data.draw(s).itemsize == 4
+
+
+@checks_deprecated_behaviour
+@given(nps.arrays(dtype="U", shape=1, elements=st.just("abc\0\0")))
+def test_unicode_string_dtype_not_trimmed(arr):
+    assert arr[0] == u"abc"
 
 
 def test_test_basic_indices_kwonly_emulation():